fix: split Tabs compoonents
