+++ packages/components/tabs/src/index.ts
@@ -1,3 +1,8 @@
 export { Tabs } from './CompoundTabs';
-export { Tab, TabList, TabPanel } from './Tabs';
-export type { TabProps, TabPanelProps, TabsProps, TabListProps } from './Tabs';
+export { Tab } from './Tab';
+export { TabList } from './TabList';
+export { TabPanel } from './TabPanel';
+export type { TabsProps } from './Tabs';
+export type { TabProps } from './Tab';
+export type { TabListProps } from './TabList';
+export type { TabPanelProps } from './TabPanel';
